moved all components to the components folder
