+++ QuizMakerAngular/src/app/app.module.ts
@@ -7,17 +7,17 @@ import { HttpClientModule } from '@angular/common/http';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { UserProfileComponent } from './userProfile/userProfile.component';
+import { UserProfileComponent } from './components/userProfile/userProfile.component';
 //import { UsernameComponent } from './username/username.component';
-import { PasswordComponent } from './password/password.component';
-import { UsernameComponent } from './username/username.component';
-import { HeaderComponent } from './header/header.component';
-import { MainComponent } from './main/main.component';
+import { PasswordComponent } from './components/password/password.component';
+import { UsernameComponent } from './components/username/username.component';
+import { HeaderComponent } from './components/header/header.component';
+import { MainComponent } from './components/main/main.component';
 import { NavComponent } from './nav/nav.component';
-import { FooterComponent } from './footer/footer.component';
-import { AsideComponent } from './aside/aside.component';
-import { PrivacyPolicyComponent } from './privacy-policy/privacy-policy.component';
-import { LandingpageComponent } from './landingpage/landingpage.component';
+import { FooterComponent } from './components/footer/footer.component';
+import { AsideComponent } from './components/aside/aside.component';
+import { PrivacyPolicyComponent } from './components/privacy-policy/privacy-policy.component';
+import { LandingpageComponent } from './components/landingpage/landingpage.component';
