+++ QuizMakerAngular/src/app/components/footer/footer.component.spec.ts
