+++ QuizMakerAngular/src/app/components/footer/footer.component.ts
