+++ QuizMakerAngular/src/app/components/privacy-policy/privacy-policy.component.html
